Changes
Added void return type to tests methods `setUp` and `tearDown` to
match `Orchestra\Testbench\TestCase::setUp()` and
`Orchestra\Testbench\TestCase::tearDown()`.

Changed the `handle` method in the `AddCustomProvider` middleware to
override the `invalidProvider` call result when no provider is given
in the case of a `client_credentials` authentication request.

Added new tests in the `AddCustomProviderTest` class to test the new
cases.

Modified existing tests in the `AddCustomProviderTest` class to define use
of the `Request` class `has` method.